batch: Move candidate fingerprints
Candidate stable-ID policy now lives in operation_candidate_fingerprints.
Candidate construction imports those functions under local names that do
not reuse the former embedded helper names.

Preview-state storage reads the algorithm version from the same fingerprint
module. That keeps state invalidation tied to the ID payload owner instead
of candidate materialization.

This commit moves buffer, ownership, target-result, scope, and batch hash
construction out of operation_candidates.

Candidate preview construction still builds the same previews, but the
module now focuses on merge enumeration and target materialization.
